refactor(UNT-T23346): code refactoring and documentation update
