+++ .npmignore
@@ -7,3 +7,4 @@ CONTRIBUTING.md
 babel.config.js
 .buckconfig
 jest-setup.js
+.husky
