+++ src/components/Waveform/WaveformStyles.ts
