+++ src/components/Waveform/WaveformTypes.ts
@@ -1,9 +1,8 @@
-import type { Ref } from 'react';
 import type { StyleProp, ViewStyle } from 'react-native';
 import type { FinishMode, PlayerState, RecorderState } from '../../constants';
 import type { IStartRecording } from '../../types';
 
-export type StaticOrLive = 'static' | 'live';
+type StaticOrLive = 'static' | 'live';
 
 interface BaseWaveform {
   candleSpace?: number;
@@ -26,31 +25,18 @@ export interface LiveWaveform extends BaseWaveform {
   onRecorderStateChange?: (recorderState: RecorderState) => void;
 }
 
-export type IWaveform<T extends StaticOrLive> = T extends 'static'
-  ? StaticWaveform
-  : LiveWaveform;
+export type IWaveform = StaticWaveform | LiveWaveform;
 
 export interface IStartPlayerRef {
   finishMode?: FinishMode;
 }
-export interface IPlayWaveformRef {
+export interface IWaveformRef {
   startPlayer: (args?: IStartPlayerRef) => Promise<boolean>;
   stopPlayer: () => Promise<boolean>;
   pausePlayer: () => Promise<boolean>;
   resumePlayer: (args?: IStartPlayerRef) => Promise<boolean>;
-}
-
-export interface IRecordWaveformRef {
   startRecord: (args?: Partial<IStartRecording>) => Promise<boolean>;
   stopRecord: () => Promise<string>;
   pauseRecord: () => Promise<boolean>;
   resumeRecord: () => Promise<boolean>;
 }
-
-export type WaveformRefProps<T extends StaticOrLive> = T extends 'static'
-  ? IPlayWaveformRef
-  : IRecordWaveformRef;
-
-export type WaveformWithStandardRef<T extends StaticOrLive> = IWaveform<T> & {
-  ref: Ref<WaveformRefProps<T>>;
-};
