+++ src/components/WaveformCandle/WaveformCandleStyles.ts
