+++ src/components/WaveformCandle/WaveformCandleTypes.ts
