--- src/constants/duration.ts
@@ -1,4 +1,0 @@
-export enum DurationType {
-  current,
-  max,
-}
