--- src/constants/eventNames.ts
@@ -1,6 +1,0 @@
-export enum NativeEvents {
-  onDidFinishPlayingAudio = 'onDidFinishPlayingAudio',
-  onCurrentDuration = 'onCurrentDuration',
-  onCurrentExtractedWaveformData = 'onCurrentExtractedWaveformData',
-  onCurrentRecordingWaveformData = 'onCurrentRecordingWaveformData',
-}
