--- src/constants/permissionStatus.ts
@@ -1,5 +1,0 @@
-export enum PermissionStatus {
-  denied = 'denied',
-  undetermined = 'undetermined',
-  granted = 'granted',
-}
