--- src/constants/playerFinishMode.ts
@@ -1,5 +1,0 @@
-export enum FinishMode {
-  loop = 0,
-  pause = 1,
-  stop = 2,
-}
