--- src/constants/playerStatus.ts
@@ -1,11 +1,0 @@
-export enum PlayerState {
-  playing = 'playing',
-  paused = 'paused',
-  stopped = 'stopped',
-}
-
-export enum RecorderState {
-  recording = 'recording',
-  paused = 'paused',
-  stopped = 'stopped',
-}
