--- src/constants/updateFrequency.ts
@@ -1,7 +1,0 @@
-//Note: If you are making change here, please make sure to make change in iOS and Android side as well other wise there will be mismatch in value
-//Use same values in iOS and Android side as well
-export enum UpdateFrequency {
-  high = 250.0,
-  medium = 500.0,
-  low = 1000.0,
-}
