+++ src/hooks/useAudioPlayer.tsx
@@ -1,5 +1,5 @@
 import { NativeEventEmitter, NativeModules } from 'react-native';
-import { AudioWaveform } from '../audioWaveform';
+import { AudioWaveform } from '../AudioWaveform';
 import { NativeEvents } from '../constants';
 import {
   type IDidFinishPlayings,
@@ -21,75 +21,56 @@ export const useAudioPlayer = () => {
     NativeModules.AudioWaveformsEventEmitter
   );
 
-  const extractWaveformData = (args: IExtractWaveform) => {
-    return AudioWaveform.extractWaveformData(args);
-  };
+  const extractWaveformData = (args: IExtractWaveform) =>
+    AudioWaveform.extractWaveformData(args);
 
-  const preparePlayer = (args: IPreparePlayer) => {
-    return AudioWaveform.preparePlayer(args);
-  };
+  const preparePlayer = (args: IPreparePlayer) =>
+    AudioWaveform.preparePlayer(args);
 
-  const playPlayer = (args: IStartPlayer) => {
-    return AudioWaveform.startPlayer(args);
-  };
+  const playPlayer = (args: IStartPlayer) => AudioWaveform.startPlayer(args);
 
-  const pausePlayer = (args: IPausePlayer) => {
-    return AudioWaveform.pausePlayer(args);
-  };
+  const pausePlayer = (args: IPausePlayer) => AudioWaveform.pausePlayer(args);
 
-  const stopPlayer = (args: IStopPlayer) => {
-    return AudioWaveform.stopPlayer(args);
-  };
+  const stopPlayer = (args: IStopPlayer) => AudioWaveform.stopPlayer(args);
 
-  const seekToPlayer = (args: ISeekPlayer) => {
-    return AudioWaveform.seekToPlayer(args);
-  };
+  const seekToPlayer = (args: ISeekPlayer) => AudioWaveform.seekToPlayer(args);
 
-  const setVolume = (args: ISetVolume) => {
-    return AudioWaveform.setVolume(args);
-  };
+  const setVolume = (args: ISetVolume) => AudioWaveform.setVolume(args);
 
-  const stopAllPlayers = () => {
-    return AudioWaveform.stopAllPlayers();
-  };
+  const stopAllPlayers = () => AudioWaveform.stopAllPlayers();
 
   const getDuration = (args: IGetDuration) => AudioWaveform.getDuration(args);
 
   const onDidFinishPlayingAudio = (
     callback: (result: IDidFinishPlayings) => void
-  ) => {
-    return audioPlayerEmitter.addListener(
+  ) =>
+    audioPlayerEmitter.addListener(
       NativeEvents.onDidFinishPlayingAudio,
       result => callback(result)
     );
-  };
 
   const onCurrentDuration = (
     callback: (result: IOnCurrentDurationChange) => void
-  ) => {
-    return audioPlayerEmitter.addListener(
-      NativeEvents.onCurrentDuration,
-      result => callback(result)
+  ) =>
+    audioPlayerEmitter.addListener(NativeEvents.onCurrentDuration, result =>
+      callback(result)
     );
-  };
 
   const onCurrentExtractedWaveformData = (
     callback: (result: IOnCurrentExtractedWaveForm) => void
-  ) => {
-    return audioPlayerEmitter.addListener(
+  ) =>
+    audioPlayerEmitter.addListener(
       NativeEvents.onCurrentExtractedWaveformData,
       result => callback(result)
     );
-  };
 
   const onCurrentRecordingWaveformData = (
     callback: (result: IOnCurrentRecordingWaveForm) => void
-  ) => {
-    return audioPlayerEmitter.addListener(
+  ) =>
+    audioPlayerEmitter.addListener(
       NativeEvents.onCurrentRecordingWaveformData,
       result => callback(result)
     );
-  };
 
   return {
     extractWaveformData,
